Build Exult as a library
The existing build structure always tries to build exult as an executable.  This ends up causing link errors when building for android.  This patch modifies the build structure to build the entire exult codebase into a library and then link that library into an executable on platforms that support it.  This will enable android builds to opt out of the executable while still building the library which the app will load at runtime.