Use SDL_RWops for unzip
The unzip code is hard-coded to to directly access the filesystem, which ends up bypassing the logic in SDL_RWops for handling nonstandard directory structures and filesystems.  This patch modifies the unzip code to use the SDL_RWops abstraction.

Note: It might be even better to use the U7 abstractions, but they would require more extensive modifications to the unzip code, so sticking with SDL_RWops for now to minimize code changes.