Add SDL dependency
Download and build the SDL library with the NDK toolchain.

Note: The java layer currently has a copy of the java sources from SDL, but it would be better if the gradle build downloaded the SDL sources and used the java pieces from there.  If that change is made, then we should also utilize that same set of SDL soures for the C++ build instead of downloading a frsh copy here.