Add a simple launcher activity
This is a first pass at a simple launcher.  It includes a button to start the game and a fixed set of checkboxes which prompt the user to browse to a downloaded archive file they want to install.

This needs more work, to improve the overall look and feel, enable installing arbitrary mods, and providing links to download content.  It also is missing error checking (if something fails to install, there is no feedback and the box remains checked), nor any sort of progress bar for slower installations.