Fix mod name extraction from path
Existing code assumes paths returned by U7ListFiles will have the same form as the path passed in, but in some cases, if a relative path is passed in, the paths it returns will be resolved to absolute pathnames.  This modifies modmgr code to extract the mode name by working backward from the end of the string rather than the current offset scheme.