Add SDL Android App Java sources
This copies in the java sources distributed with SDL for building Android apps.  This is a stop-gap solution, and less than ideal as these could get out of sync with the underlying SDL library and will require manual maintenance.  These need to be kept in sync with the SDL release being pulled in by the NDK build of exult driven by cmake.

Ideally, I'd like to find a way to download these in the gradle build (possibly using [source dependencies](https://blog.gradle.org/introducing-source-dependencies)), but the SDL repository isn't set up to easily support this.