Add autolaunch checkbox
This adds a checkbox to the launcher which is off by default, but can be turned on so that starting the app will bypass the launcher UI and go directly into Exult.  Users can always exit the Exult game engine to get back to the launcher UI if they want to make changes or disable autolaunch.