Add classes for reading Exult content from archives
Setting up Exult on a desktop typically involves manually installing original game files, optional audio packs, and optional mods, and editing exult.cfg to point to their locations.  This gets particularly awkward for mobile devices.  To ease the process, the Android version of Exult includes a launcher which allows the user to browse to downloaded content archives and it rummages around inside them to find the correct content to install into the correct game folder.  This patch adds the classes that implement the archive extraction, validation, and installation.  A subsequent patch will implement the UI which drives the process.

Note that this is currently implemented in Java and only applicable to Android, but could (and probably should) be re-implemented in the C++ layer so that other platforms could make use of it as well.