Use SDL_RWops for file access where possible
Some platforms have nontraditional filesystem access and SDL has already provided an abstraction layer for them in the form of [SDL_RWops](https://wiki.libsdl.org/SDL_RWops).  Rather than introduce platform-specific code into exult to account for platform filesystem differences, leverage SDL's abstraction as the primary interface for interacting with the filesystem.

Note that there is one glaring deficiency in `SDL_RWops`: it does not provide an api for interacting with directories.  So while this change works for reading and writing individual files, it does not help with searching directories to see what they contain, so we end up falling back to legacy code that bypasses SDL, which might not work on all platforms.  This might cast doubt on whether there is value in switching to `SDL_RWops`, but the only place exult currently queries directories is when detecting mods.  This means mods need to live in a traditional filesystem location, but all other content can be placed in platform-specific locations.  This works out well for android where some of the data files can be bundled directly into the apk, while mods are installed into the device's filesystem where they can be detected.