Fix out-of-tree builds
Existing cflags include a relative path to the `headers` include directory.  When trying to build out-of-tree, the build fails to find that path and fails.  This patch modifies cflags to look for `headers` in the source tree.

Note: while this change is enough to get basic build to work out-of-tree, there are problems with some of the internal tools which prevent building mods out-of-tree.