Fix ogg/vorbis stuttering on Android
The existing 20Hz buffering for ogg/vorbis playback isn't sufficient on android.  It results in stuttering when using digital music on both an emulated android device on a laptop and on two native android devices.  Increasing to 1Hz buffering,  Note: I haven't extensively tested to find the minimum needed for these Android devices, so this could be tuned down to a smaller value if there are memory concerns.