Add a configure option for building an Android APK
This adds a new `--enable-android-apk` option to the configure script which searches for android build tools and compiles and Android APK.  This patch is only concerned with the top-level tools detection and launching the gradle build.  Subsequent patches will extend the Android app to pull in the exult sources and related dependencies.