Fix crash when restarting game on Android
This probably doesn't matter for standalone executables that fully terminate the process when the game exits, but on Android, the game is loaded as a shared library and it is not reliably unloaded when the game exits.  In cases where the library is kept in memory, a subsequent attempt to start the game will find the previous run's `SDL_Window` still allocated, and if we don't clean it up here, the subsequent attempt to call `SDL_CreateWindow()` will crash.